gemini-webrtc-web-simple/server/requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[daily,silero,google,openai]
